fix to allow tables to converted to HTML

+++ convert.py
@@ -1,17 +1,20 @@
+import sys
 import os
 import mammoth
 from mammoth.cli import ImageWriter
+import html2markdown
 import requests
 import re
 import datetime
 import urllib.parse
 from slugify import slugify
+from bs4 import BeautifulSoup as bs
 
-def main():
+def main(filename):
     # erase() # uncomment to erase existing content
     
     # open list of urls to convert
-    with open('urls.txt', 'r') as f:
+    with open(filename, 'r') as f:
         for line in f:
             process(line.strip())
 
@@ -19,10 +22,14 @@ def erase():
     folders = ['images', 'markdown', 'docx']
     for folder in folders:
         for file in os.listdir(folder):
+            # if file is .gitkeep or .gitignore, skip it
+            if file == '.gitkeep' or file == '.gitignore':
+                continue
             os.remove(folder + '/' + file)
 
 def process(url):
-    url = url.replace('/edit', '/export?format=docx')
+    # clean url
+    url = url.replace(url.split('/')[-1], '') + 'export?format=docx'
 
     # get document
     r = requests.get(url)
@@ -31,7 +38,7 @@ def process(url):
     filename = urllib.parse.unquote(r.headers['Content-Disposition'].split('filename=')[1].split(';')[1].replace("filename*=UTF-8''", ''))
 
     # title is filename without extension
-    title = filename.split('.')[0]
+    title = filename.split('.')[0].strip()
 
     # slug is the filename without the extension, slugified
     slug = slugify(filename.split('.')[0])
@@ -43,10 +50,11 @@ def process(url):
     #  using mammoth, convert the docx to markdown
     with open('docx/' + filename, 'rb') as f:
         # process with mammoth
-        result = mammoth.convert_to_markdown(f, convert_image=mammoth.images.img_element(ImageWriter('images')))
+        result = mammoth.convert_to_html(f, convert_image=mammoth.images.img_element(ImageWriter('images')))
 
         # markdown is the result of the conversion
-        md = result.value
+        # md = result.value
+        md = html2markdown.convert(result.value)
 
         # remove all ids and classes from the markdown
         md = re.sub(r'\s+id="[^"]*"', '', md)
@@ -57,6 +65,9 @@ def process(url):
 
         # remove any escaped periods
         md = re.sub(r'\\.', '.', md)
+
+        # remove any &nbsp;
+        md = re.sub(r'&nbsp;', '', md)
 
         # get todays date
         date = datetime.datetime.now().strftime("%Y-%m-%d")
@@ -95,12 +106,67 @@ description: {description}
 
             # replace the image in the markdown with the new image name
             # change the base_url to wherever your images are stored
-            base_url = '/images/'
+            base_url = '/static/img/blog/'
             md = md.replace(image, base_url + new_name)
+
+        # find all tables in the markdown
+        tables = re.findall(r'<table>.*</table>', md, re.DOTALL)
+
+        # loop tables
+        for table in tables:
+            # transform table
+            md = md.replace(table, transform_table(table))
 
     # save the markdown to the /markdown folder
     with open('markdown/' + slug + '.md', 'w') as f:
         f.write(md)
 
+def transform_table(table):
+    # wrap table in a div with class table-expand
+    table = '<div class="table-expand">\n' + table + '\n</div>'
+    
+    # add classes table is-bordered is-narrow is-fullwidth mb-3 to table element
+    table = re.sub(r'<table>', '<table class="table is-bordered is-narrow is-fullwidth mb-3">', table)
+
+    # prettify table using BeautifulSoup
+    soup = bs(table, 'html.parser')
+    
+    # remove any <p> or </p> tags from th elements
+    for th in soup.find_all('th'):
+        th_text = th.text
+        p = th.find('p')
+        if p:
+            p.extract()
+        th.string = th_text
+    
+    # do same for td elements
+    for td in soup.find_all('td'):
+        td_text = td.text
+        p = td.find('p')
+        if p:
+            p.extract()
+        td.string = td_text
+
+    table = soup.prettify()
+
+    # return table
+    return table
+
 if __name__ == '__main__':
-    main()
+    # if -r is passed as an argument, run erase()
+    if '-r' in sys.argv:
+        erase()
+
+    # default file is urls.txt
+    file = 'urls.txt'
+
+    # get -u to specify list of urls to convert
+    if '-u' in sys.argv:
+        file = sys.argv[sys.argv.index('-u') + 1]
+
+    # if file does not exist, exit
+    if not os.path.exists(file):
+        print(f'List of urls does not exist. Make sure the {file} file is created first.')
+        exit()
+
+    main(file)
